refactor(build-test): relocate BuildTestApp and BuildTestDialog files to standard directories
- Move BuildTestApp.mjs to scripts/apps/documents/dialog/
- Move BuildTestDialog.html to templates/documents/tests/
- Update relative import paths and Handlebars template path references accordingly
